requirements.txt
# HTTP server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# HuggingFace models (free, open-source, auto-download from hub)
transformers==4.44.2
# CPU-only torch — installed explicitly in Dockerfile
# On GPU: pip install torch (auto-selects CUDA)
torch
sentencepiece

# Image processing
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==11.0.0

# Utilities
python-dotenv==1.0.1
